Compare # advisories ignored to advisory ID
Previously this did an overall check of totals, which might get us into
a weird spot if a CVE was removed and dependencies were added (though
unlikely). This makes a more specific comparison so that we can be sure
each potential problem is evaluated. It also adds some output that can
help with understanding why the task failed.